% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_plots.R
\name{plot_variable_heatmap}
\alias{plot_variable_heatmap}
\title{Visualize Variable Co-occurrence Heatmap}
\usage{
plot_variable_heatmap(
  plausible_result,
  col_palette = c("white", "lightblue", "blue", "darkblue")
)
}
\arguments{
\item{plausible_result}{Output from plausible_models()}

\item{col_palette}{Color palette for heatmap. Default: blue gradient.
Can be a vector of colors or a single color name.}
}
\value{
A heatmap plot
}
\description{
Creates a heatmap showing how often pairs of variables appear together
in the plausible models. This reveals variable relationships and redundancy.
}
