% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_paths.R
\name{print.multipath}
\alias{print.multipath}
\title{Print method for multipath objects}
\usage{
\method{print}{multipath}(x, ...)
}
\arguments{
\item{x}{A multipath object from build_paths()}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for multipath objects
}
