% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_models.R
\name{print.plausible_models}
\alias{print.plausible_models}
\title{Print method for plausible_models objects}
\usage{
\method{print}{plausible_models}(x, ...)
}
\arguments{
\item{x}{A plausible_models data frame}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for plausible_models objects
}
