% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{print.stability}
\alias{print.stability}
\title{Print method for stability objects}
\usage{
\method{print}{stability}(x, ...)
}
\arguments{
\item{x}{A stability object from stability()}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for stability objects
}
