% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print_confusion_matrix}
\alias{print_confusion_matrix}
\title{Print Confusion Matrix Nicely}
\usage{
print_confusion_matrix(conf_mat)
}
\arguments{
\item{conf_mat}{Confusion matrix from confusion_metrics()}
}
\description{
Print Confusion Matrix Nicely
}
